requirements.txt
pandas
mysql-connector-python
boto3
python-dotenv
